Add Price and MonthlyRent Specifications and Refactor SpecificationBuilder

+++ src/main/java/com/example/PropertyDemo/Property/RentalProperty.java
@@ -10,21 +10,21 @@ import java.util.List;
 @Entity
 public class RentalProperty extends Property {
 
-    private double monthlyRent;
+    private int monthlyRent;
 
     public RentalProperty() {
     }
 
-    public RentalProperty(PropertyType type, Location location, int bedrooms, List<URL> images, Agent agent, double monthlyRent) {
+    public RentalProperty(PropertyType type, Location location, int bedrooms, List<URL> images, Agent agent, int monthlyRent) {
         super(type, location, bedrooms, images, agent);
         this.monthlyRent = monthlyRent;
     }
 
-    public double getMonthlyRent() {
+    public int getMonthlyRent() {
         return monthlyRent;
     }
 
-    public void setMonthlyRent(double monthlyRent) {
+    public void setMonthlyRent(int monthlyRent) {
         this.monthlyRent = monthlyRent;
     }
