+++ src/test/java/com/example/PropertyDemo/Builders/RentalPropertyBuilder.java
@@ -3,6 +3,6 @@ package com.example.PropertyDemo.Builders;
 import com.example.PropertyDemo.Property.RentalProperty;
 
 public interface RentalPropertyBuilder<T extends RentalPropertyBuilder<T>> extends PropertyBuilder<T>{
-    T withMonthlyRent(double monthlyRent);
+    T withMonthlyRent(int monthlyRent);
     RentalProperty build();
 }
